backend/twitter/requirements.txt
psycopg2-binary>=2.9.0
twikit>=2.0.0